Move more things around (mostly into interfaces.py)
